Conflicts:
	include/ContentManager.h
	include/DX11Renderer.h
	src/ContentManager.cpp
	src/DX11Renderer.cpp
	src/TestGame.cpp

+++ tool/Development/src/Mesh.cpp
@@ -172,9 +172,9 @@ void MeshData::Export(std::string path, Handle handle)
 	fileData.vertexMeta.elementSize = sizeof(Export::Mesh::Vertex);
 
 	fileData.indexMeta.offsetInBytes = fileData.vertexMeta.offsetInBytes + fileData.vertexMeta.sizeInBytes;
-	fileData.indexMeta.sizeInBytes = (uint16_t)(sizeof(uint16_t) * handle->faceCount);
+	fileData.indexMeta.sizeInBytes = (uint16_t)(sizeof(uint32_t) * handle->faceCount);
 	fileData.indexMeta.elementCount = (uint16_t)handle->faceCount;
-	fileData.indexMeta.elementSize = sizeof(uint16_t);
+	fileData.indexMeta.elementSize = sizeof(uint32_t);
 
 	fwrite(&fileData.blockCount, sizeof(uint8_t), 1, meshFile);
 	fwrite(&fileData.boundsMeta.offsetInBytes, sizeof(Export::Mesh::Block), fileData.blockCount, meshFile);
